+++ CHANGES.md
@@ -8,13 +8,14 @@
 
 ##### Additions :tada:
 
+- Added new constructors to `LocalHorizontalCoordinateSystem` taking ECEF<->Local transformation matrices directly.
 - Unknown properties in objects read with a `JsonReader` are now stored in the `unknownProperties` property on `ExtensibleObject` by default. To ignore them, as was done in previous versions, call `setCaptureUnknownProperties` on `JsonReaderOptions`.
-
-### v0.27.0 - 2023-09-01
 
 ##### Fixes :wrench:
 
 - Fixed a bug where an empty error message would get propagated to a tileset's `loadErrorCallback`.
+- Fixed several small build script issues to allow cesium-native to be used in Univeral Windows Platform (UWP) applications, such as those that run on Holo Lens 2.
+- When KTX transcoding fails, the image will now be fully decompressed instead of returning an error.
 
 ### v0.26.0 - 2023-08-01